another changes on Supplier tests and Create ClientService

+++ hypermarket-core/src/main/java/ro/sda/hypermarket/core/service/SupplierServiceImpl.java
@@ -29,23 +29,35 @@ public class SupplierServiceImpl implements SupplierService {
     }
 
     @Override
-    public Supplier getSupplierById(Long id) {
-        return supplierDao.getSupplierById(id);
+    public Supplier getSupplierById(Long id, boolean useHibernate) {
+        if (useHibernate) {
+            return supplierDao.getSupplierById(id);
+        }
+        return supplierRepository.findById(id);
     }
 
     @Override
-    public List<Supplier> getAllSuppliers() {
+    public List<Supplier> getAllSuppliers(boolean useHibernate) {
+        if (useHibernate) {
             return supplierDao.getAllSuppliers();
+        }
+        return supplierRepository.findAll();
     }
 
     @Override
-    public void deleteSupplier(Supplier supplier) {
+    public void deleteSupplier(Supplier supplier, boolean useHibernate) {
+        if (useHibernate) {
         supplierDao.deleteSupplier(supplier);
+        }
+        supplierRepository.delete(supplier);
     }
 
     @Override
-    public void updateSupplier(Supplier supplier) {
+    public void updateSupplier(Supplier supplier, boolean useHibernate) {
+        if (useHibernate) {
         supplierDao.updateSupplier(supplier);
+        }
+        supplierRepository.save(supplier);
     }
 
 }
